% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignRel.R
\name{assignRel}
\alias{assignRel}
\title{Assigns the type of relationships for each pair that is above the selected relatedness threshold.}
\usage{
assignRel(kinshipFile, t = 0.0125)
}
\arguments{
\item{kinshipFile}{plink format pairwise relatedness table. This can be obtrained from the prel function.}

\item{t}{pi_hat threshold for pairwise relationships}
}
\value{
This script will output a plain txt file for the related pairs and the type of relationships the pair has.
}
\description{
Part of the pretrel package. This functions asssigns the type of relationship for each pair above the selected relatedness threshold. The input file needs to be in the plink pairwise relatedness format.
}
\seealso{
selectSNPs(), inputPrepare(), pret(), prel(), pretrel()
}
\author{
Salih Tuna, \email{st620@cam.ac.uk}
}
