% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excludeSamples_1000G}
\alias{excludeSamples_1000G}
\title{Samples to remove from the 1000G dataset as these samples found to be related.}
\format{A plain text file with list of samples to be excluded from the analysis}
\usage{
excludeSamples_1000G
}
\description{
Samples to remove from the 1000G dataset as these samples found to be related.
}
\keyword{datasets}
