% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputPrepare.R
\name{inputPrepare}
\alias{inputPrepare}
\title{Prepares the input VCF file for ethnicity and relatedness analysis by using selected set of SNPs.}
\usage{
inputPrepare(snpList, inputRAW_VCF, RefDataDir, nnode = 2)
}
\arguments{
\item{snpList}{List of SNPs that will be used to filter the inputRAW and reference set. By default this is the output file from selectSNPs() function. You can enter a list of snps (ex 1    12345)}

\item{inputRAW_VCF}{unfiltered vcf file(s) to be used for the analysis}

\item{RefDataDir}{Path for the reference set}

\item{nnode}{number of nodes to be used for the parallel job execution. Default is two nodes.}
}
\value{
This script will output a merged VCF file that contains both project and reference samples. The merged VCF file consists of only the SNPs selected either by the selectSNPs function or the user input.
}
\description{
InputPrepare function prepares the input VCF file for ethnicity and relatedness analysis by extracting the set of SNPs from the reference set and the set of samples and merge the two together. It is a series of bcftools commands and some of which run in parallel. So it relies on bcftools and GNU Parallel softwares. These two softwares need to be installed and be in your path. Please see 'References' section below where to get these softwares. The function makes sure that exact same of SNPs exist in both the test samples and the reference set. Currently this runs only on 1000Genome reference set(see references for details).
}
\references{
\url{https://samtools.github.io/bcftools/bcftools.html}

\url{https://www.gnu.org/software/parallel/}

\url{https://www.internationalgenome.org/data/}
}
\seealso{
selectSNPs(), pret(), prel(), pretrel()
}
\author{
Salih Tuna, \email{st620@cam.ac.uk}
}
