% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prel.R
\name{prel}
\alias{prel}
\title{prel is used for predicting the pairwise relationships.}
\usage{
prel(inputVCF = "../mergedInput.vcf.gz", maxUnrelSet = TRUE,
  t = 0.0125)
}
\arguments{
\item{inputVCF}{A vcf file with the selected set of SNPs and which includes samples of interest together with the reference set of samples}

\item{maxUnrelSet}{Logical, A list of maximum unrelated set will be calculated using PRIMUS. This can be turned off by setting it to FALSE. This will call the Primus(https://primus.gs.washington.edu/primusweb/) software. Make sure that Primus is installed and it is in your path}

\item{t}{pi_hat threshold for pairwise relationships}
}
\value{
This script will output the predicted ethnicities based on reference set used (PopAssigned.txt), pca plot in pdf format, eigenvalues and eigenvectors as a plain text format, and the pairwise relatedness.
}
\description{
prel is a function for predicting pairwise relatedness using the GENESIS package. Furthermore, this function outputs a plain file with the type of relationship for each pair that is above the selected relatedness threshold (e.g. Sample1 sample2 Parent/Offspring). Using PRIMUS (see references) software, it outputs the the list of maximum unrelated set of samples. This function can be used on its own or as part of the pretrel package.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/GENESIS.html}

\url{https://primus.gs.washington.edu/primusweb/}
}
\seealso{
selectSNPs(), inputPrepare(), pret(), pretrel(), assignRel()
}
\author{
Salih Tuna, \email{st620@cam.ac.uk}
}
