% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pret.R
\name{pret}
\alias{pret}
\title{PRedicting EThnicity and RELatedness (pretrel) from genotyping data.}
\usage{
pret(inputVCF)
}
\arguments{
\item{inputVCF}{Merged vcf file that has samples from both reference and the project set with the selected set of SNPs. This file can be obtained by running the inputPrepare() function of the pretrel package.}
}
\value{
This script will output PopAssigned.txt, for each sample showing the predicted ethnicity, pca plot in pdf format, eigenvalues and eigenvectors as a plain text format.
}
\description{
pret is used for predicting the ethnicity of the samples of interest. It uses Principal Component Analysis (PCA). GENESIS package is used as the base for appliying pca to a vcf file. Once the projections are obtained by pc-air function from GENESIS package, samples are assigned to one of the predefined clusters, from the reference set.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/GENESIS.html}
}
\seealso{
selectSNPs(), inputPrepare(), prel(), pretrel()
}
\author{
Salih Tuna, \email{st620@cam.ac.uk}
}
