% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrel.R
\name{pretrel}
\alias{pretrel}
\title{PRedicting EThnicity and RELatedness (pretrel) from genotyping data.}
\usage{
pretrel(snpList, inputRAW_VCF, RefDataDir, inputVCF)
}
\arguments{
\item{snpList}{List of SNPs that will be used to filter the inputRAW and reference set}

\item{inputRAW_VCF}{unfiltered vcf file(s) to be used for the analysis}

\item{RefDataDir}{Path to the 1000G reference set (ALL.chr*.phase3_shapeit2_mvncall_integrated_v5a.20130502.genotypes.vcf.gz)}

\item{inputVCF}{A vcf file with the selected set of SNPs and which includes samples of interest together with the reference set of samples}
}
\value{
This script will output the predicted ethnicities based on reference set used (PopAssigned.txt), pca plot in pdf format, eigenvalues and eigenvectors as a plain text format, and the pairwise relatedness.
}
\description{
pretrel function is a wrapper for four functions that starts with selectSNP() function for selecting SNPs that will be used throughout the analysis. If the user would like to user their own list of SNPs, then this selectSNPs() function can be ignored. Next function within pretrel is to run inputPrepare() which will prepare the input file using the reference set of samples, snp list and the raw vcf file. Next step is to predict the ethnicities(pret()) and predict the pairwise relatedness and also selects the set of maximum unrelated samples (prel()). All these functions can be used seperately.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/GENESIS.html}

\url{https://primus.gs.washington.edu/primusweb/}
}
\seealso{
selectSNPs(), inputPrepare(), pret(), prel()
}
