% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs.R
\name{selectSNPs}
\alias{selectSNPs}
\title{Selects SNPs that will be used for the ethnicity and relatedness analysis}
\usage{
selectSNPs(inputRAW_VCF, RefDataDir, tmpDir, nnode = 2)
}
\arguments{
\item{inputRAW_VCF}{unfiltered vcf file(s) to be used for the analysis.}

\item{RefDataDir}{Path for the reference set.}

\item{tmpDir}{Directory to write the temporary files.}

\item{nnode}{number of nodes to run the jobs in parallel.}
}
\value{
This script will output a list of SNPs in plain text format. This file will be used as input for the inputPrepare() function.
}
\description{
selectSNPs is a series of bcftools commands. So it relies on bcftools and GNU Parallel. These two softwares need to be installed and be in your path. Please see below where to get these softwares. SelectSNPs function can be used on its own or as part of the pretrel() function. This approach requires that both the reference set of samples and project set of samples are mapped to the same reference version (i.e. either both is v37 or both is v38).
}
\references{
\url{https://samtools.github.io/bcftools/bcftools.html}

\url{https://www.gnu.org/software/parallel/}
}
\author{
Salih Tuna, \email{st620@cam.ac.uk}
}
